--- src/tests/fetcher_test.rs
@@ -1,4 +1,0 @@
-#[test]
-fn proxy_test() {
-    todo!();
-}
