--- src/tests/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(test)]
-mod parser_test;
-
-#[cfg(test)]
-mod fetcher_test;
